Resolved conflicts:
- pyproject.toml: Keep both playwright and networkx dependencies
- docs/user-guide/examples/sigenergy-system.md: Accept main version (better formatting)
- uv.lock: Regenerated with all dependencies

+++ custom_components/haeo/data/loader/extractors/open_meteo_solar_forecast.py
@@ -50,9 +50,9 @@ class Parser:
         return all(is_parsable_to_datetime(k) and isinstance(v, (int, float)) for k, v in watts.items())
 
     @staticmethod
-    def extract(state: OpenMeteoSolarState) -> tuple[Sequence[tuple[float, float]], str, SensorDeviceClass]:
+    def extract(state: OpenMeteoSolarState) -> tuple[Sequence[tuple[int, float]], str, SensorDeviceClass]:
         """Extract forecast data from Open-Meteo solar forecast format."""
-        parsed: list[tuple[float, float]] = [
+        parsed: list[tuple[int, float]] = [
             (parse_datetime_to_timestamp(time), value) for time, value in state.attributes["watts"].items()
         ]
         parsed.sort(key=lambda x: x[0])
